+++ packages/moengage_cards/moengage_cards_android/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MoEngage Cards Android Plugin
           
+# Next Release Date
+
+## Next Version
+- Add jsonEncode for missing APIs
+- `cards-core` version updated to `2.3.1`
+
 # 25-11-2024
 
 ## 4.0.2